Download point-level ensemble weather forecasting using open-meteo API

Download point-level ensemble weather forecasting using open-meteo API

Usage

get_forecast(
  latitude,
  longitude,
  site_id = NULL,
  forecast_days,
  past_days,
  model = "generic",
  variables = c("temperature_2m")
)

Arguments

latitude

latitude degree north

longitude

longitude degree east

site_id

name of site location (optional, default = NULL)

forecast_days

Number of days in the future for forecast (starts at current day)

past_days

Number of days in the past to include in the data

model

id of forest model https://open-meteo.com/en/docs/climate-api. Default = "generic"

variables

vector of name of variable(s) https://open-meteo.com/en/docs/ensemble-api.

Value

data frame (in long format)

Examples

if (FALSE) { # interactive()
get_forecast(latitude = 37.30,
            longitude = -79.83,
            forecast_days = 7,
            past_days = 2,
           model = "generic",
            variables = c("temperature_2m"))
}